Move retry_on_exception into its own file
