+++ metaspace/engine/sm/engine/util.py
@@ -1,10 +1,7 @@
 import logging
 import re
-from functools import wraps
 import os
 from pathlib import Path
-import random
-from time import sleep
 
 from sm.engine.config import init_loggers, SMConfig
 
@@ -67,29 +64,3 @@ class GlobalInit:
 
     def __exit__(self, ext_type, ext_value, traceback):
         self.pool.close()
-
-
-def retry_on_exception(exception_type=Exception, num_retries=3, retry_wait_params=(2, 3, 5)):
-    def decorator(func):
-        func_name = getattr(func, '__name__', 'Function')
-
-        @wraps(func)
-        def wrapper(*args, **kwargs):
-            for i in range(num_retries):
-                try:
-                    return func(*args, **kwargs)
-                except exception_type as e:
-                    wait_initial, wait_increase, jitter = retry_wait_params
-                    min_wait = wait_initial + i * wait_increase
-                    delay = random.uniform(min_wait, min_wait + jitter)
-                    logger.warning(
-                        f'{func_name} raised {type(e)} on attempt {i+1}. '
-                        f'Retrying after {delay:.1f} seconds...'
-                    )
-                    sleep(delay)
-            # Last attempt, don't catch the exception
-            return func(*args, **kwargs)
-
-        return wrapper
-
-    return decorator
